perf: faster replace find first of with find (#232)
* src: remove std::find from set_host_or_hostname + while loop in get_host_delimiter_location

* Not bad.

* Good enough.

* Temporary commit.

* Cleaning.

* Cleaning.

* SWAR implementation.

* Fix for big endian systems.

* Adding swap_bytes_if_big_endian

* Fixed BE function.

---------

+++ src/checkers.cpp
@@ -47,14 +47,14 @@ namespace ada::checkers {
                     path_signature_table[uint8_t(input[i + 7])]);
     }
     for (; i < input.size(); i++) {
-      accumulator |= path_signature_table[uint8_t(input[i])];
+      accumulator |= uint8_t(path_signature_table[uint8_t(input[i])]);
     }
     return accumulator;
   }
 
 
   ada_really_inline constexpr bool verify_dns_length(std::string_view input) noexcept {
-    if(input.back() == '.') { 
+    if(input.back() == '.') {
       if(input.size() > 254) return false;
     } else if (input.size() > 253) return false;
